pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openapi-chroma"
version = "0.1.0"
description = "OpenAPI spec ingestion into ChromaDB with MCP server for Claude"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "jsonref>=1.1.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
openapi-chroma = "main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]
